aws/requirements.txt
# aws cdk
aws-cdk-lib==2.23.0
constructs==10.1.5

# pydantic settings
pydantic
python-dotenv
